Extract `ShareAnnotations` component
